refactor: 💡 Upgrade discord and add bot test run script
