+++ src/utils/discord/battleNotifier/bnCommand/bnBattleTypes.js
