+++ src/utils/discord/battleNotifier/bnCommand/config.js
