+++ src/utils/discord/battleNotifier/bnCommand/subCommands/getBn.js
