+++ src/utils/discord/battleNotifier/bnCommand/subCommands/helpBn.js
