+++ src/utils/discord/battleNotifier/bnCommand/subCommands/setBn.js
