+++ src/utils/discord/battleNotifier/bnCommand/subCommands/testBn.js
