+++ src/utils/discord/battleNotifier/bnCommand/subCommands/toggleBn.js
