+++ src/utils/discord/battleNotifier/bnCommand/userConfig/userConfigFormatter.js
