+++ src/utils/discord/battleNotifier/bnCommand/userConfig/userConfigParser.js
