+++ src/utils/discord/battleNotifier/notifyBattle.js
