+++ src/utils/discord/discord.js
@@ -1,8 +1,8 @@
-import Discord from 'discord.js';
-import moment from 'moment';
-import { forEach } from 'lodash';
-import config from '../config';
-import createBN from './battleNotifier';
+const Discord = require('discord.js');
+const moment = require('moment');
+const { forEach } = require('lodash');
+const config = require('../../config');
+const createBN = require('./battleNotifier');
 
 const client = new Discord.Client();
 const battleNotifier = createBN({
@@ -10,7 +10,7 @@ const battleNotifier = createBN({
   client,
 });
 
-export function discord() {
+function discord() {
   client.once('ready', () => {
     client.user.setPresence({
       status: 'online',
@@ -39,9 +39,9 @@ export function discord() {
   });
 }
 
-export function sendMessage(channel, message) {
+function sendMessage(channel, message) {
   if (config.discord.token) {
-    client.channels.get(channel).send(message);
+    client.channels.cache.get(channel).send(message);
   }
 }
 
@@ -127,7 +127,7 @@ const cripple = content => {
   return '';
 };
 
-export function discordChatline(content) {
+function discordChatline(content) {
   const ts = moment(content.timestamp, 'YYYY-MM-DD HH:mm:ss UTC').format(
     'HH:mm:ss',
   );
@@ -137,7 +137,7 @@ export function discordChatline(content) {
   );
 }
 
-export function discordBesttime(content) {
+function discordBesttime(content) {
   if (!content.battleIndex) {
     let text = `${formatLevel(content.level)}:`;
     text += ` ${content.time} by ${content.kuski} (${content.position}.)`;
@@ -148,7 +148,7 @@ export function discordBesttime(content) {
   }
 }
 
-export function discordBestmultitime(content) {
+function discordBestmultitime(content) {
   if (!content.battleIndex) {
     sendMessage(
       config.discord.channels.times,
@@ -168,7 +168,7 @@ const battleToString = battle => {
   return text;
 };
 
-export async function discordBattlestart(content) {
+async function discordBattlestart(content) {
   const battleString = battleToString(content);
   const text = `${config.discord.icons.started} **${battleString}`;
   sendMessage(config.discord.channels.battle, text);
@@ -176,7 +176,7 @@ export async function discordBattlestart(content) {
   battleNotifier.notifyBattle(content, battleString);
 }
 
-export function discordBattlequeue(content) {
+function discordBattlequeue(content) {
   if (content.queue.length > 0) {
     let text = `${config.discord.icons.queue} **Queue:`;
     content.queue.map(q => {
@@ -194,7 +194,7 @@ export function discordBattlequeue(content) {
   }
 }
 
-export function discordBattleEnd(content) {
+function discordBattleEnd(content) {
   if (content.aborted) {
     let text = `${config.discord.icons.ended} **`;
     text += `${battleIn(content.battleType, content.level)}${cripple(
@@ -204,7 +204,7 @@ export function discordBattleEnd(content) {
   }
 }
 
-export function discordBattleresults(content) {
+function discordBattleresults(content) {
   let text = `${config.discord.icons.results} **`;
   text += battleIn(content.battleType, content.level);
   text += `${cripple(content)} by ${content.designer} over**\n`;
@@ -252,3 +252,15 @@ client.on('message', async message => {
     message.reply('There was an error trying to execute that command!');
   }
 });
+
+module.exports = {
+  discord,
+  sendMessage,
+  discordChatline,
+  discordBesttime,
+  discordBestmultitime,
+  discordBattlestart,
+  discordBattlequeue,
+  discordBattleEnd,
+  discordBattleresults,
+};
