+++ src/utils/discord/discordTester.js
@@ -1,0 +1,3 @@
+const { discord } = require('./discord');
+
+discord();
